Add cmake to dev.rake; Make build of vim-spf13 conditional

+++ tasks/vim.rake
@@ -1,6 +1,6 @@
 namespace :vim do
   COLORSCHEMES = {
-    "clouds-midnight" => "http://darcsden.com/alex/dotfiles/raw/.vim/colors/clouds-midnight.vim"
+    :clouds_midnight => "http://darcsden.com/alex/dotfiles/raw/.vim/colors/clouds-midnight.vim"
   }
 
   def install_colorscheme(cs=nil)
@@ -25,12 +25,14 @@ namespace :vim do
   #desc "Install vim-spf13"
   task :vim_spf13 => ["vim"] do
     FileUtils.cd(HOME) do
-      sh("curl http://j.mp/spf13-vim3 -L -o - | sh") unless File.exists?(".vimrc-bundles")
+      unless File.exists?(".spf13-vim-3") do
+          sh("curl http://j.mp/spf13-vim3 -L -o - | sh") unless File.exists?(".vimrc-bundles")
 
-      # Install clouds-mignight colorscheme
-      install_colorscheme('clouds-midnight')
+          # Install clouds-mignight COLORSCHEMES
+          install_colorscheme(:clouds_midnight)
 
-      notice("Installed vim spf13")
+          notice("Installed vim spf13")
+      end
     end
   end
 end
